now doing kernel timings with cuda events

+++ include/cuda_mesh_operations.hpp
@@ -7,10 +7,10 @@
 #include <timing_struct.hpp>
 
 namespace ab {
-	void normals_by_area_weight_he_cuda(HalfedgeMesh* mesh, size_t threads = 256,size_t blocks=1, timing_struct& timing=timing_struct());
-	void normals_by_area_weight_sm_cuda(SimpleMesh* mesh, size_t threads = 256, size_t blocks = 1, timing_struct& timing = timing_struct());
-	void centroids_he_cuda(HalfedgeMesh* mesh, std::vector<float3>& centroids_array, size_t threads = 256,size_t blocks = 1, timing_struct& timing = timing_struct());
-	void centroids_sm_cuda(SimpleMesh* mesh, std::vector<float3>& centroids, size_t threads = 256, size_t blocks = 1, timing_struct& timing = timing_struct());
+	void normals_by_area_weight_he_cuda(HalfedgeMesh* mesh, int threads = 256,int blocks=1, timing_struct& timing=timing_struct());
+	void normals_by_area_weight_sm_cuda(SimpleMesh* mesh, int threads = 256, int blocks = 1, timing_struct& timing = timing_struct());
+	void centroids_he_cuda(HalfedgeMesh* mesh, std::vector<float3>& centroids_array, int threads = 256,int blocks = 1, timing_struct& timing = timing_struct());
+	void centroids_sm_cuda(SimpleMesh* mesh, std::vector<float3>& centroids, int threads = 256, int blocks = 1, timing_struct& timing = timing_struct());
 
 	//float3 extensions
 	__forceinline__ __host__ __device__ float3 cross3df(const float3 a, const float3 b) {
